migrate microdata action tests
